requirements.txt
requests>=2.32
PyYAML>=6
